Now testing with the livejournal graph
Testing with array reuse in vprog. This is didn't give correct answers when I wrote Jose's algorithm. See if it's true still

New in this version is that we use one element double array to transmit page rank

Adding program like PR with Jose's graphs -- now let's create the graph with array vertex. in this version they get replaced with primitives, but at least once we have arrays